feat(maquette): one-click per-agent config button in TopBar (T26)
Extract AgentModal (create/manage agent dialog) from FleetDashboard into
its own shared file (agents/AgentModal.tsx), exporting AgentModal +
AgentModalMode + MODELS/MODEL_META/slugify; onFlash is now optional.
TopBar gains a Settings2 'Agent configuration' button left of the global
Settings gear (in-agent only) that opens the same Manage dialog for the
active agent, collapsing the four-step switcher journey to one click. The
modal is rendered as a sibling of the .vibrancy header (fragment return)
so its absolute-inset backdrop anchors to the viewport.

chore: sync tree descriptions
